remove unused css
