+++ frontend/src/Components/Auth/Authenticator.jsx
@@ -40,12 +40,6 @@ const useStyles = makeStyles((theme) => ({
     flexDirection: 'row',
     justifyContent: 'space-between',
   },
-  form: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: theme.spacing(2),
-    padding: theme.spacing(0, 2),
-  },
 }));
 
 const Authenticator = () => {
