+++ frontend/src/Components/CommunityEvent/ReportCommunityEvent.jsx
@@ -11,6 +11,7 @@ import {
   Checkbox,
 } from '@material-ui/core';
 
+import dayjs from 'dayjs';
 import { produce } from 'immer';
 import { enqueueSnackbar } from 'notistack';
 import { useDispatch, useSelector } from 'react-redux';
@@ -18,7 +19,6 @@ import { useDispatch, useSelector } from 'react-redux';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import { eventActions } from '../../Containers/Event/eventSlice';
 import { BLANK_REPORT_FORM, BLANK_REPORT_FORM_ERROR, BLANK_REPORT_FORM_TOUCHED } from '../Organization/constants';
-import dayjs from 'dayjs';
 
 const useStyles = makeStyles((theme) => ({
   container: {
