+++ frontend/src/Components/CommunityEvent/constants.js
@@ -1,3 +1,32 @@
+export const SKILLS_REQUIRED_OPTIONS = [
+  'Event Planning',
+  'Volunteer Coordination',
+  'Fundraising',
+  'Public Speaking',
+  'Social Media Management',
+  'Graphic Design',
+  'Photography',
+  'Videography',
+  'Community Outreach',
+  'Project Management',
+  'Writing',
+  'Marketing',
+  'Event Promotion',
+  'Logistics',
+  'Catering',
+  'Decorations',
+  'Tech Support',
+  'Sponsorship',
+  'First Aid',
+  'Education',
+  'Entertainment',
+  'Environmental Conservation',
+  'Art and Crafts',
+  'Music',
+  'Dance',
+  'Other',
+];
+
 export const BLANK_NEW_EVENT = {
   title: '',
   cause: '',
