+++ frontend/src/Components/Event/SearchAllEvents.jsx
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from 'react';
-import { Autocomplete } from '@material-ui/lab';
-import { FiberManualRecordRounded } from '@material-ui/icons';
 import { makeStyles } from '@material-ui/core/styles';
+import { FiberManualRecordRounded } from '@material-ui/icons';
 import { Box, TextField, Typography } from '@material-ui/core';
+
 import classNames from 'classnames';
+import { Autocomplete } from '@material-ui/lab';
 
 const useStyles = makeStyles((theme) => ({
   text: {
