+++ frontend/src/Components/Event/constants.jsx
@@ -1,7 +1,9 @@
 import { Chip, Tooltip } from '@material-ui/core';
 import { AssignmentTurnedInRounded, GroupWorkRounded } from '@material-ui/icons';
+
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
+
 dayjs.extend(relativeTime);
 
 export const UNKNOWN_LABEL = 'Unknown';
