+++ frontend/src/Components/Home/HomePageHeaderMap.jsx
@@ -1,4 +1,6 @@
 import React, { useEffect, useRef } from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+
 import './map.css';
 import Map from 'ol/Map.js';
 import View from 'ol/View.js';
@@ -17,7 +19,6 @@ import LayerTile from 'ol/layer/Tile.js';
 import SourceOSM from 'ol/source/OSM.js';
 import PopupContent from './PopupContent';
 import { Attribution, defaults as defaultControls } from 'ol/control.js';
-import { makeStyles } from '@material-ui/core/styles';
 
 const center = proj.fromLonLat([-92.0369, 38.9072]);
