+++ frontend/src/Components/ViewEventListCta/NavigateProfileFromEventCta.jsx
@@ -13,11 +13,6 @@ const useStyles = makeStyles((theme) => ({
   text: {
     fontSize: '0.925rem',
   },
-  spinnerContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'center',
-  },
 }));
 
 const NavigateProfileFromEventCta = () => {
