+++ frontend/src/Components/ViewExpenseList/ViewExpenseList.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import EmptyComponent from '../../util/EmptyComponent';
 import { makeStyles } from '@material-ui/core/styles';
 import { CircularProgress, Container } from '@material-ui/core';
 import { CancelRounded, DoneRounded } from '@material-ui/icons';
-import EasyEdit, { Types } from 'react-easy-edit';
-import { eventActions } from '../../Containers/Event/eventSlice';
+
 import dayjs from 'dayjs';
-import relativeTime from 'dayjs/plugin/relativeTime';
 import List from '../DrawerListComponent/List';
+import EasyEdit, { Types } from 'react-easy-edit';
+import relativeTime from 'dayjs/plugin/relativeTime';
+import { useDispatch, useSelector } from 'react-redux';
+import EmptyComponent from '../../util/EmptyComponent';
 import { VIEW_EXPENSE_LIST_COLUMN_HEADERS } from './constants';
 
 const useStyles = makeStyles((theme) => ({
